컬럼 CRUD 추가 수정

+++ services/column.service.js
@@ -8,9 +8,9 @@ class ColumnService {
   columnRepository = new ColumnRepository();
 
   // 컬럼 추가
-  createColumn = async (boardId, name, position) => {
+  createColumn = async (boardId, name, position, userId) => {
     if (isNaN(boardId) || boardId < 1) {
-      throw new MakeError(400, '잘못된 boardId입니다.');
+      throw new MakeError(400, '잘못된 boardId 형식입니다.');
     }
 
     if (!name) {
@@ -25,7 +25,15 @@ class ColumnService {
       boardId,
     );
     if (!findBoardData) {
-      throw new MakeError(400, '존재하지 않는 보드입니다.');
+      throw new MakeError(404, '존재하지 않는 보드입니다.');
+    }
+
+    const findCollaborator = await this.columnRepository.checkCollaborator(
+      boardId,
+      userId,
+    );
+    if (!findCollaborator) {
+      throw new MakeError(403, '해당 보드에 초대된 회원이 아닙니다.');
     }
 
     await this.columnRepository.createColumn(boardId, name, position);
@@ -33,19 +41,31 @@ class ColumnService {
   };
 
   // 컬럼 조회
-  getAllColumns = async (boardId) => {
+  getAllColumns = async (boardId, userId) => {
+    if (isNaN(boardId) || boardId < 1) {
+      throw new MakeError(400, '잘못된 boardId 형식입니다.');
+    }
+
     const findBoardData = await this.columnRepository.findOneBoardDataByBoardId(
       boardId,
     );
     if (!findBoardData) {
-      throw new MakeError(400, '존재하지 않는 보드입니다.');
+      throw new MakeError(404, '존재하지 않는 보드입니다.');
+    }
+
+    const findCollaborator = await this.columnRepository.checkCollaborator(
+      boardId,
+      userId,
+    );
+    if (!findCollaborator) {
+      throw new MakeError(403, '해당 보드에 초대된 회원이 아닙니다.');
     }
 
     return await this.columnRepository.getAllColumns(boardId);
   };
 
   // 컬럼 수정(name)
-  modifyNameOfColumn = async (boardId, columnId, name) => {
+  modifyNameOfColumn = async (boardId, columnId, name, userId) => {
     if (isNaN(boardId) || boardId < 1) {
       throw new MakeError(400, '잘못된 boardId 형식입니다.');
     }
@@ -59,9 +79,20 @@ class ColumnService {
     }
 
     const findColumnData =
-      await this.columnRepository.findOneColumnDataByCondition({ columnId });
+      await this.columnRepository.findOneColumnDataByCondition({
+        boardId,
+        columnId,
+      });
     if (!findColumnData) {
-      throw new MakeError(400, '존재하지 않는 컬럼입니다.');
+      throw new MakeError(404, '존재하지 않는 컬럼입니다.');
+    }
+
+    const findCollaborator = await this.columnRepository.checkCollaborator(
+      boardId,
+      userId,
+    );
+    if (!findCollaborator) {
+      throw new MakeError(403, '해당 보드에 초대된 회원이 아닙니다.');
     }
 
     await this.columnRepository.modifyNameOfColumn(columnId, name);
@@ -69,7 +100,7 @@ class ColumnService {
   };
 
   // 컬럼 수정(position)
-  modifyPositionOfColumn = async (boardId, columnId, position) => {
+  modifyPositionOfColumn = async (boardId, columnId, position, userId) => {
     if (isNaN(boardId) || boardId < 1) {
       throw new MakeError(400, '잘못된 boardId 형식입니다.');
     }
@@ -88,12 +119,21 @@ class ColumnService {
 
     try {
       const findColumnData =
-        await this.columnRepository.findOneColumnDataByCondition({ columnId });
+        await this.columnRepository.findOneColumnDataByCondition({
+          boardId,
+          columnId,
+        });
       if (!findColumnData) {
-        throw new MakeError(400, '존재하지 않는 컬럼입니다.');
+        throw new MakeError(404, '존재하지 않는 컬럼입니다.');
       }
 
-      const originalPosition = findColumnData.position;
+      const findCollaborator = await this.columnRepository.checkCollaborator(
+        boardId,
+        userId,
+      );
+      if (!findCollaborator) {
+        throw new MakeError(403, '해당 보드에 초대된 회원이 아닙니다.');
+      }
 
       const columnWithNewPosition =
         await this.columnRepository.findOneColumnDataByCondition({
@@ -107,8 +147,11 @@ class ColumnService {
         transaction,
       );
 
+      const originalPosition = findColumnData.position;
+      const columnWithNewPositionColumnId = columnWithNewPosition.columnId;
+
       await this.columnRepository.modifyPositionOfColumn(
-        columnWithNewPosition.columnId,
+        columnWithNewPositionColumnId,
         originalPosition,
         transaction,
       );
@@ -123,7 +166,7 @@ class ColumnService {
   };
 
   // 컬럼 삭제
-  deleteColumn = async (boardId, columnId) => {
+  deleteColumn = async (boardId, columnId, userId) => {
     if (isNaN(boardId) || boardId < 1) {
       throw new MakeError(400, '잘못된 boardId 형식입니다.');
     }
@@ -133,9 +176,20 @@ class ColumnService {
     }
 
     const findColumnData =
-      await this.columnRepository.findOneColumnDataByCondition({ columnId });
+      await this.columnRepository.findOneColumnDataByCondition({
+        boardId,
+        columnId,
+      });
     if (!findColumnData) {
-      throw new MakeError(400, '존재하지 않는 컬럼입니다.');
+      throw new MakeError(404, '존재하지 않는 컬럼입니다.');
+    }
+
+    const findCollaborator = await this.columnRepository.checkCollaborator(
+      boardId,
+      userId,
+    );
+    if (!findCollaborator) {
+      throw new MakeError(403, '해당 보드에 초대된 회원이 아닙니다.');
     }
 
     await this.columnRepository.deleteColumn(columnId);
